pyproject.toml
[project]
name = "multi-vendor-digital-store"
version = "0.1.0"
description = "Digital Marketplace is a multi-vendor e-commerce platform built with Django, designed for selling digital products such as e-books, music, software, and more. The platform allows multiple vendors to register, list their digital products, and manage sales, while customers can browse, purchase, and download products instantly."
authors = [
    {name = "Amir",email = "amirbehvandi747@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-phonenumber-field (>=8.0.0,<9.0.0)",
    "django-phonenumbers (>=1.0.1,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
